+++ postinstall.sh
@@ -1,48 +1,51 @@
 #!/bin/sh
 
-# Bashscript which is executed by bash *AFTER* complete installation is done
-# (but *BEFORE* postupdate). Use with caution and remember, that all systems
-# may be different! Better to do this in your own Pluginscript if possible.
-#
-# Exit code must be 0 if executed successfull.
-#
-# Will be executed as user "loxberry".
-#
-# We add 5 arguments when executing the script:
-# command <TEMPFOLDER> <NAME> <FOLDER> <VERSION> <BASEFOLDER>
+# Runs as loxberry after the plugin files have been copied.
 
-ARGV2=$2 # Second argument is Plugin-Name for scipts etc.
-ARGV3=$3 # Third argument is Plugin installation folder
-ARGV5=$5 # Fifth argument is Base folder of LoxBerry
+PTEMPDIR=$1
+PSHNAME=$2
+PDIR=$3
+PVERSION=$4
+PTEMPPATH=$6
 
-/bin/sed -i "s#REPLACEBYSUBFOLDER#$ARGV3#" $ARGV5/config/plugins/$ARGV3/smartmeter.cfg
-/bin/sed -i "s#REPLACEBYNAME#$ARGV2#" $ARGV5/config/plugins/$ARGV3/smartmeter.cfg
-/bin/sed -i "s#REPLACELBHOMEDIR#$ARGV5#" $ARGV5/bin/plugins/$ARGV3/reboot_cron_runner.sh
-/bin/sed -i "s#REPLACELBPPLUGINDIR#$ARGV3#" $ARGV5/bin/plugins/$ARGV3/reboot_cron_runner.sh
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_config.pl
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_validate.pl
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_control.pl
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/vzlogger_mqtt_bridge.pl
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/smartmeter_legacy_runtime.pl
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/install_vzlogger_bridge_service.sh
-/bin/chmod +x $ARGV5/bin/plugins/$ARGV3/install_vzlogger_service_override.sh
-/bin/chmod +x $ARGV5/webfrontend/htmlauth/plugins/$ARGV3/vzlogger_live.cgi
-/bin/chmod +x $ARGV5/webfrontend/htmlauth/plugins/$ARGV3/vzlogger_config.cgi
+for required in LBHOMEDIR LBPCONFIG LBPBIN LBPCGI; do
+	eval "value=\${$required:-}"
+	if [ -z "$value" ]; then
+		echo "<ERROR> Required LoxBerry V4 environment variable $required is missing."
+		exit 2
+	fi
+done
+
+PCONFIG="$LBPCONFIG/$PDIR"
+PBIN="$LBPBIN/$PDIR"
+PCGI="$LBPCGI/$PDIR"
+
+/bin/sed -i "s#REPLACEBYSUBFOLDER#$PDIR#" "$PCONFIG/smartmeter.cfg"
+/bin/sed -i "s#REPLACEBYNAME#$PSHNAME#" "$PCONFIG/smartmeter.cfg"
+/bin/sed -i "s#REPLACELBHOMEDIR#$LBHOMEDIR#" "$PBIN/reboot_cron_runner.sh"
+/bin/sed -i "s#REPLACELBPPLUGINDIR#$PDIR#" "$PBIN/reboot_cron_runner.sh"
+
+for executable in \
+	"$PBIN/vzlogger_config.pl" \
+	"$PBIN/vzlogger_validate.pl" \
+	"$PBIN/vzlogger_control.pl" \
+	"$PBIN/vzlogger_mqtt_bridge.pl" \
+	"$PBIN/smartmeter_legacy_runtime.pl" \
+	"$PCGI/vzlogger_live.cgi" \
+	"$PCGI/vzlogger_config.cgi"
+do
+	/bin/chmod 0755 "$executable"
+done
 
 echo "<INFO> Rename htaccess to .htaccess"
-mv $ARGV5/webfrontend/htmlauth/plugins/$ARGV3/htaccess $ARGV5/webfrontend/htmlauth/plugins/$ARGV3/.htaccess
+mv "$PCGI/htaccess" "$PCGI/.htaccess"
 
 echo "<INFO> vzLogger package is installed through LoxBerry dpkg/apt dependencies."
-
 if command -v mosquitto_sub >/dev/null 2>&1; then
 	echo "<INFO> mosquitto_sub found for MQTT bridge."
 else
 	echo "<WARNING> mosquitto_sub is not available. Install mosquitto-clients for HTTP/UDP cache updates from vzLogger MQTT."
 fi
 
-echo "<INFO> *******************************************************************"
-echo "<INFO> * Please reboot your LoxBerry to initialize the Smartmeter v2 Plugin *"
-echo "<INFO> *******************************************************************"
-
-# Exit with Status 0
+echo "<OK> SmartMeter v2 files prepared. No reboot is required."
 exit 0